// src/app/interfaces/economia.interface.ts
// Generated by https://quicktype.io

export interface IEconomia {
    status:       string;
    totalResults: number;
    results:      Result[];
    nextPage:     string;
    showFullContent?: boolean;
}

export interface Result {
    showFullContent?: boolean;
    title:       string;
    link:        string;
    keywords:    Keyword[] | null;
    creator:     string[] | null;
    video_url:   null;
    description: null | string;
    content:     string;
    pubDate:     string;
    image_url:   string;
    source_id:   SourceID;
    category:    Category[];
    country:     Country[];
    language:    Language;
    expanded?: boolean;
}

export enum Category {
    Business = "business",
}

export enum Country {
    Spain = "spain",
}

export enum Keyword {
    EconomíaFinanzas = "Economía Finanzas",
    Finanzas = "Finanzas",
}

export enum Language {
    Spanish = "spanish",
}

export enum SourceID {
    Esinvesting = "esinvesting",
    Europapress = "europapress",
}